Move Reporter base class into event module
and several small code improvements
